updated: data module directories
