--- core/data/src/main/java/dev/enesky/core/data/Images.kt
@@ -1,39 +1,0 @@
-/*
- *                          Copyright 2023
- *            Designed and developed by Enes Kamil Yılmaz
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package dev.enesky.core.data
-
-import android.os.Parcelable
-import com.google.gson.annotations.SerializedName
-import kotlinx.parcelize.Parcelize
-
-/**
- * Created by Enes Kamil YILMAZ on 28/10/2023
- */
-@Parcelize
-data class Images(
-    val jpg: ImageList? = null,
-    val webp: ImageList? = null,
-) : Parcelable
-
-@Parcelize
-data class ImageList(
-    @SerializedName("image_url") val imageUrl: String? = null,
-    @SerializedName("small_image_url") val smallImageUrl: String? = null,
-    @SerializedName("medium_image_url") val mediumImageUrl: String? = null,
-    @SerializedName("large_image_url") val largeImageUrl: String? = null,
-    @SerializedName("maximum_image_url") val maximumImageUrl: String? = null,
-) : Parcelable
