+++ core/data/src/main/java/dev/enesky/core/data/models/Character.kt
@@ -14,15 +14,34 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.enesky.core.data
+package dev.enesky.core.data.models
+
+import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
+import kotlinx.parcelize.Parcelize
 
 /**
  * Created by Enes Kamil YILMAZ on 28/10/2023
  */
 
-enum class AnimeFilter(val filter: String) {
-    AIRING("airing"),
-    UPCOMING("upcoming"),
-    POPULARITY("bypopularity"),
-    FAVORITE("favorite"),
-}
+@Parcelize
+data class Character(
+    val character: Person,
+    val role: String,
+    val favorites: Int,
+    val voiceActors: List<VoiceActor>,
+) : Parcelable
+
+@Parcelize
+data class Person(
+    @SerializedName("mal_id") val id: Int,
+    val url: String,
+    val images: Images,
+    val name: String,
+) : Parcelable
+
+@Parcelize
+data class VoiceActor(
+    val person: Person,
+    val language: String,
+) : Parcelable
